requirements.txt
# Core Dependencies
python-opcua>=1.0.0
matplotlib>=3.5.0
cryptography>=3.4.8
openpyxl>=3.0.9
weasyprint>=56.0

# Optional Dependencies (for enhanced features)
pymodbus>=3.0.0  # For advanced Modbus support
snap7>=1.3  # For S7 PLC communication
pycomm3>=1.2.0  # For Allen-Bradley communication

# Development Dependencies (optional)
pytest>=7.0.0  # For testing
black>=22.0.0  # For code formatting
flake8>=4.0.0  # For linting
